# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: ReST-Race App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
paths:
  /user:
    x-swagger-router-controller: user
    get:
      operationId: list
      description: Returns a list of users.
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: save
      description: Add a new user.
      parameters:
        - name: Properties
          description: User properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{name}:
    x-swagger-router-controller: user
    get:
      operationId: get
      description: Get a user.
      parameters:
      - name: name
        type: string
        in: path
        required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetUserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: Update a user.
      parameters:
      - name: name
        type: string
        in: path
        required: true
      - name: email
        description: User properties
        in: body
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: remove
      description: Remove a user.
      parameters:
      - name: name
        description: User name
        type: string
        in: path
        required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  User:
    type: object
    properties:
      password:
        type: string
        description: user password
      email:
        type: string
        description: user email
      name:
        type: string
        description: user name
    required:
      - name
      - email
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  GetUserResponse:
    required:
      - name
      - email
    properties:
      name:
        type: string
      email:
        type: string
  GetUserListResponse:
    required:
      - users
    properties:
      users:
        type: array
        items:
          $ref: '#/definitions/User'

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string